requirements.txt
fastapi
python-dateutil
cbpro
pydantic
uvicorn
a2wsgi